P3 follow-up: move occ/ under cc/ to match the other protocols
occ/ was the only CC implementation still sitting at the repo root after
the rest moved to cc/. It uses a plain Makefile on the legacy uint64-key
API and is not wired into CMake, so it's a cosmetic move — not a
functional change — but worth doing so cc/ contains every protocol.

- git mv occ -> cc/occ
- Relative includes shifted to account for the extra depth:
    .cc files at cc/occ/:        "../include/..."   -> "../../include/..."
    headers at cc/occ/include/:  "../../include/..." -> "../../../include/..."
- Makefile relative paths adjusted similarly (REL, -I, masstree include,
  clean target).
- Doc references in CLAUDE.md and docs/protocols.md updated.

The Makefile still references include/MakefileForMasstreeUse which is
not in the tree, so the build was already broken before this move; not
in scope to fix here. Porting cc/occ/ to ccbench_add_protocol(...) is
the natural follow-up.
