+++ cc/occ/include/atomic_tool.hh
@@ -2,7 +2,7 @@
 
 #include "common.hh"
 
-#include "../../include/inline.hh"
+#include "../../../include/inline.hh"
 
 INLINE uint64_t atomicLoadGE();
