+++ cc/occ/include/common.hh
@@ -6,9 +6,9 @@
 
 #include "tuple.hh"
 
-#include "../../include/cache_line_size.hh"
-#include "../../include/int64byte.hh"
-#include "../../include/masstree_wrapper.hh"
+#include "../../../include/cache_line_size.hh"
+#include "../../../include/int64byte.hh"
+#include "../../../include/masstree_wrapper.hh"
 
 #include "gflags/gflags.h"
 #include "glog/logging.h"
