+++ cc/occ/occ.cc
@@ -19,18 +19,18 @@
 #include "include/transaction.hh"
 #include "include/util.hh"
 
-#include "../include/atomic_wrapper.hh"
-#include "../include/backoff.hh"
-#include "../include/cpu.hh"
-#include "../include/debug.hh"
-#include "../include/fileio.hh"
-#include "../include/masstree_wrapper.hh"
-#include "../include/random.hh"
-#include "../include/result.hh"
-#include "../include/rwlock.hh"
-#include "../include/tsc.hh"
-#include "../include/util.hh"
-#include "../include/zipf.hh"
+#include "../../include/atomic_wrapper.hh"
+#include "../../include/backoff.hh"
+#include "../../include/cpu.hh"
+#include "../../include/debug.hh"
+#include "../../include/fileio.hh"
+#include "../../include/masstree_wrapper.hh"
+#include "../../include/random.hh"
+#include "../../include/result.hh"
+#include "../../include/rwlock.hh"
+#include "../../include/tsc.hh"
+#include "../../include/util.hh"
+#include "../../include/zipf.hh"
 
 using namespace std;
