+++ cc/occ/result.cc
@@ -1,8 +1,8 @@
 #include "include/result.hh"
 #include "include/common.hh"
 
-#include "../include/cache_line_size.hh"
-#include "../include/result.hh"
+#include "../../include/cache_line_size.hh"
+#include "../../include/result.hh"
 
 using namespace std;
