+++ cc/occ/transaction.cc
@@ -12,12 +12,12 @@
 #include "include/log.hh"
 #include "include/transaction.hh"
 
-#include "../include/backoff.hh"
-#include "../include/debug.hh"
-#include "../include/fileio.hh"
-#include "../include/masstree_wrapper.hh"
-#include "../include/tsc.hh"
-#include "../include/util.hh"
+#include "../../include/backoff.hh"
+#include "../../include/debug.hh"
+#include "../../include/fileio.hh"
+#include "../../include/masstree_wrapper.hh"
+#include "../../include/tsc.hh"
+#include "../../include/util.hh"
 
 extern void displayDB();
